renamed server folder to .server as per remix documentation - https://remix.run/docs/en/main/discussion/server-vs-client
